refactor(omni): use native Megatron integration
