+++ scripts/models/qwen3_omni_moe.sh
@@ -1,53 +1,6 @@
-NLAYERS=48
-FIRST_K_DENSE_REPLACE=0
+source "$(cd -- "$(dirname -- "${BASH_SOURCE[0]}")" && pwd)/qwen3-30B-A3B.sh"
 
-arr=()
-for ((i=0; i<NLAYERS; i++)); do
-  if (( i < FIRST_K_DENSE_REPLACE )); then
-    arr+=(0)
-  else
-    arr+=(1)
-  fi
-done
-
-printf -v MOE_LAYER_FREQ "[%s]" "$(IFS=', '; echo "${arr[*]}")"
-
-
-# Qwen3-Omni-30B-A3B Thinker (text + vision + audio). Same MoE core as Qwen3-30B-A3B.
-MODEL_ARGS=(
-   --disable-bias-linear
-   --qk-layernorm
-   --group-query-attention
-   --num-attention-heads 32
-   --num-query-groups 4
-   --kv-channels 128
-   --num-layers 48
-   --hidden-size 2048
-   --ffn-hidden-size 6144
-
-   --normalization RMSNorm
-   --position-embedding-type rope
-   --norm-epsilon 1e-6
-   --rotary-percent 1.0
-   --swiglu
-   --untie-embeddings-and-output-weights
-   --vocab-size 151936
-
-   --rotary-base "${MODEL_ARGS_ROTARY_BASE:-1000000}"
-
-   # moe
-   --moe-ffn-hidden-size 768
-   --moe-router-score-function softmax
-   --moe-token-dispatcher-type alltoall
-   --moe-router-topk 8
-   --moe-layer-freq "$MOE_LAYER_FREQ"
-   --num-experts 128
-   --moe-grouped-gemm
-   --moe-token-drop-policy probs
-   --moe-router-dtype fp32
-   --moe-permute-fusion
-   --moe-aux-loss-coeff 0
-
-   # multimodal bridge
+MODEL_ARGS+=(
+   --spec vime_plugins.models.qwen3_omni_moe get_qwen3_omni_model_provider
    --use-qwen-vl
 )
