+++ vime/rollout/vllm_rollout.py
@@ -25,10 +25,8 @@ from vime.utils.eval_config import EvalDatasetConfig
 from vime.utils.http_utils import get, get_rollout_num_engines, post
 from vime.utils.misc import SingletonMeta, load_function
 from vime.utils.processing_utils import (
+    build_multimodal_messages,
     build_processor_kwargs,
-    encode_audio_for_rollout_engine,
-    encode_image_for_rollout_engine,
-    encode_video_for_rollout_engine,
     load_processor,
     load_tokenizer,
 )
@@ -363,11 +361,7 @@ async def generate(args: Namespace, sample: Sample, sampling_params: dict[str, A
 
     inference_sampling_params = _build_inference_sampling_params(sampling_params)
 
-    mm_inputs = sample.multimodal_inputs or {}
-    images = mm_inputs.get("images")
-    videos = mm_inputs.get("videos")
-    audios = mm_inputs.get("audio")
-    has_mm_media = bool(images) or bool(videos) or bool(audios)
+    messages = build_multimodal_messages(sample.prompt, sample.multimodal_inputs)
 
     if not sample.tokens:
         sample.tokens = prompt_ids
@@ -378,22 +372,10 @@ async def generate(args: Namespace, sample: Sample, sampling_params: dict[str, A
         if getattr(args, "router_policy", None) == "consistent_hash":
             headers = {"x-session-id": sample.session_id}
 
-    # Prepare payload for vLLM server. Any image/audio/video must go through
-    # /render so vLLM runs the matching encoders (and optional feature dumps).
-    if has_mm_media:
-        content: list[dict[str, Any]] = [{"type": "text", "text": sample.prompt}]
-        for image in images or []:
-            data_url = encode_image_for_rollout_engine(image)
-            content.append({"type": "image_url", "image_url": {"url": data_url}})
-        for audio in audios or []:
-            audio_url = encode_audio_for_rollout_engine(audio)
-            content.append({"type": "audio_url", "audio_url": {"url": audio_url}})
-        for video in videos or []:
-            video_url = encode_video_for_rollout_engine(video)
-            content.append({"type": "video_url", "video_url": {"url": video_url}})
+    if messages:
         render_payload = {
             "model": args.hf_checkpoint,
-            "messages": [{"role": "user", "content": content}],
+            "messages": messages,
         }
         await prime_encoder(args, render_payload["messages"])
         render_url = f"{base}/v1/chat/completions/render"
