--- vime_plugins/megatron_bridge/__init__.py
@@ -1,1 +1,0 @@
-import vime_plugins.megatron_bridge.qwen3_omni_moe  # noqa: F401  # register Qwen3-Omni-MoE bridge
